Rename files, add hover
