--- src/lib/component/partial/Annotation/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Annotation";
