+++ src/lib/component/partial/ExpandableKeyValueChip/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ExpandableKeyValueChip";
