+++ src/lib/component/partial/KeyValueChip/index.js
@@ -1,0 +1,1 @@
+export { default } from "./KeyValueChip";
