--- src/lib/component/partial/Label/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Label";
